Parsing dates as UTC string to get consistent parsing results
As moment parses dates using locale as default, different results are
obtained depending on where the execution happens. As gcloud functions
maight be located in different zones, the result may differ because of
that. To avoid this variation, the date is parsed as if where a UTC
date and then the Argentina offset (-03:00) is applied to the date
obtained before serializing the response.